requirements.txt
torch>=1.9.0
torchvision>=0.10.0
streamlit>=1.10.0
Pillow>=8.3.1
numpy>=1.19.5
librosa>=0.8.1
scipy>=1.7.1
soundfile>=0.10.3
matplotlib>=3.5.0
pandas>=1.3.0
scikit-learn>=1.0.0
requests>=2.27.1 